refactor: :recycle: jobs,streams refactoring
