+++ ui/src/modules/jobs/pages/Jobs.tsx
@@ -3,11 +3,10 @@ import { Button, Tabs, Empty, message, Spin } from "antd"
 import { useNavigate } from "react-router-dom"
 import { useAppStore } from "../../../store"
 import JobTable from "../components/JobTable"
-import FirstJob from "../../../assets/FirstJob.svg"
-import JobsTutorial from "../../../assets/JobsTutorial.svg"
-import { GitCommit, PlayCircle, Plus } from "@phosphor-icons/react"
+import { GitCommit, Plus } from "@phosphor-icons/react"
 import DeleteJobModal from "../../common/Modals/DeleteJobModal"
 import { jobService } from "../../../api"
+import JobEmptyState from "../components/JobEmptyState"
 
 const Jobs: React.FC = () => {
 	const [activeTab, setActiveTab] = useState("active")
@@ -158,47 +157,7 @@ const Jobs: React.FC = () => {
 							/>
 						</div>
 					) : tab.key === "active" && showEmpty ? (
-						<div className="flex flex-col items-center justify-center py-16">
-							<img
-								src={FirstJob}
-								alt="Empty state"
-								className="mb-8 h-64 w-96"
-							/>
-							<div className="mb-2 text-[#193AE6]">Welcome User !</div>
-							<h2 className="mb-2 text-2xl font-bold">
-								Ready to run your first Job
-							</h2>
-							<p className="mb-8 text-[#0A0A0A]">
-								Get started and experience the speed of O<b>Lake</b> by running
-								jobs
-							</p>
-							<Button
-								type="primary"
-								className="mb-12 bg-[#193AE6] text-sm"
-								onClick={handleCreateJob}
-							>
-								<GitCommit />
-								Create your first Job
-							</Button>
-							<div className="w-[412px] rounded-xl border-[1px] border-[#D9D9D9] bg-white p-6 shadow-sm">
-								<div className="flex items-center gap-4">
-									<img
-										src={JobsTutorial}
-										alt="Job Tutorial"
-										className="rounded-lg"
-									/>
-									<div className="flex-1">
-										<div className="mb-1 flex items-center gap-1 text-xs">
-											<PlayCircle color="#9f9f9f" />
-											<span className="text-[#9F9F9F]">OLake/ Tutorial</span>
-										</div>
-										<div className="text-xs">
-											Checkout this tutorial, to know more about running jobs
-										</div>
-									</div>
-								</div>
-							</div>
-						</div>
+						<JobEmptyState handleCreateJob={handleCreateJob} />
 					) : filteredJobs.length === 0 ? (
 						<Empty
 							image={Empty.PRESENTED_IMAGE_SIMPLE}
