+++ ui/src/types/authTypes.ts
@@ -2,3 +2,8 @@ export interface LoginArgs {
 	username: string
 	password: string
 }
+
+export interface LoginResponse {
+	username: string
+	password: string
+}
